refactor: extract shared transfer utilities into _transfer_utils.py
move non-CLI-specific helpers (collect_resources, find_root_resources,
execute_transfer, get_resource_display_name) into a shared module so
both typer and cyclopts implementations import from the same place.
